Create README.md
Rentier - an application to evaluate the final cost
of flat to buy. User can add a flat, estimate condition and furnishings of apartment. In second section
user can make a profile for furnishing flat with prices and links to offers. Then user can add renovation offer. All of that things can be combined in last part and will evaluate how many money is needed to buy a flat including all costs and taxes. To make that I’ve used MVC pattern and Spring framework. Views are made with simple Bootstrap elements. For the model,
I’ve used Hibernate entities to store data in MySQL database. I'll move project to SpringBoot
and use SpringSecure to identify users.